Practica3_PSI-main/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
# attrs==23.1.0
# autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
oauthlib==3.2.2
# ubuntu 20
psycopg2-binary==2.9.7
# psycopg2
pip==23.1.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
python-chess
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
setuptools==67.7.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
wheel==0.40.0
zope.interface==6.0
whitenoise==5.2.0
websockets==11.0.3

